change int to uint64

+++ kernel/syscall.c
@@ -101,7 +101,7 @@ extern uint64 sys_unlink(void);
 extern uint64 sys_link(void);
 extern uint64 sys_mkdir(void);
 extern uint64 sys_close(void);
-extern int sys_getcourseno(void);
+extern uint64 sys_getcourseno(void);
 
 // An array mapping syscall numbers from syscall.h
 // to the function that handles the system call.
